Start on Question 5
My plan is to write a script that runs each of the requirements individually first. Before putting them together and allowing it to run when opened. After running the functions required for each requirement, store the result. Print the result at the end of the script to show the statistics as a whole.